+++ module/src/lib.rs
@@ -2,17 +2,24 @@ use serde::{Deserialize, Serialize};
 use serde_json::Value;
 use std::collections::HashMap;
 use std::error::Error;
+use std::fmt::Display;
 use thiserror::Error;
 
 const DESCRIPTION: &str = "description";
 
 // A identifier string. Must be unique within it's context.
-#[derive(Debug, Eq, Hash, PartialEq, Deserialize, Clone, PartialOrd, Ord, Serialize)]
+#[derive(Debug, Eq, Hash, PartialEq, Clone, PartialOrd, Ord, Serialize, Deserialize)]
 pub struct Identifier(String);
 
 impl From<String> for Identifier {
     fn from(value: String) -> Self {
         Identifier(value)
+    }
+}
+
+impl Display for Identifier {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        write!(f, "{}", self.0)
     }
 }
 
@@ -32,3 +39,4 @@ pub enum ModuleError {
 
 pub mod module;
 pub mod module_type;
+pub mod rendering;