+++ module/src/rendering/helpers/split_lines.rs
@@ -1,0 +1,3 @@
+use handlebars::handlebars_helper;
+
+handlebars_helper!(split_lines: |paragraph: String| paragraph.split('\n').collect::<Vec<&str>>());